requirements.txt
cec==0.2.7
rpi-rf==0.9.7
Flask==1.1.2
pyyaml==5.3.1
cec==0.2.7